requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Scientific Computing
numpy==1.24.3
matplotlib==3.7.2
scipy==1.11.2

# Data Processing
pandas==2.0.3

# File Handling
python-dotenv==1.0.0

# Development Tools
flask-cors==4.0.0

# Optional: For advanced features
# Database (if needed)
# Flask-SQLAlchemy==3.0.5
# Flask-Migrate==4.0.5

# Background Tasks (if needed)
# Celery==5.3.1
# Redis==4.6.0

# Testing
pytest==7.4.0
pytest-flask==1.2.0

# Deployment
gunicorn==21.2.0